Web/requirements.txt
certifi==2019.11.28
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
matplotlib>=1.4.3
pandas>=0.19
tpot>=0.11.0
seaborn==0.9.0
cloudpickle>=0.6.0
scikit-learn==0.22.0
joblib==0.14.0
google-cloud-storage==1.30.0